+++ src/fhir/R4R5-common/sdc-import.js
@@ -199,6 +199,7 @@ function addSDCImportFns(ns) {
           score = !score ? LForms.Util.findObjectInArray(option.extension, 'url', self.argonautExtUrlExtensionScore) : score;
           if(score) {
             answer.score = parseFloat(score.valueDecimal);
+            answer._scoreExtUrl = score.url;
           }
 
         }