+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [41.1.1] 2026-04-02
+### Fixed
+- Original score extensions (ordinalValue or itemWeight) were not
+  kept in the exported questionnaire.
+
 ## [41.1.0] 2026-03-30
 ### Added
 - Support for rendering-markdown extension. Note that rendering-xhtml